requirements.txt
fastapi
uvicorn
pydantic
streamlit
requests
openai
sentence-transformers
faiss-cpu
python-dotenv
pandas